run.sh: always lint all files in repo, regardless of branch
Previously, we were trying to detect if we were on a branch other than master and only lint changed
files. This was based on the incorrect assumption that Drone, our CI server, ran tests on master -
it actually checks out a specific commit. This caused no files to be linted on Drone, which was the
opposite of what we wanted.

Instead, we should always lint all files. This is a stricter stance, but also simpler.